Changing int to bigint

+++ Enterprise/Subsystem/ProductLauncher/Database/Identity/Person/Stored Procedures/ListPersons_Export.sql
@@ -198,7 +198,7 @@ BEGIN
   INSERT INTO @tblPrimaryProperties (        
    Properties        
   )        
-  SELECT CONVERT(int, value)        
+  SELECT CONVERT(bigint, value)        
   FROM STRING_SPLIT(@PrimaryProperties, ',');        
  END
